requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudbiolinux==0.1
cryptography==39.0.0
fabric==3.0.0
Flask==2.2.3
Flask-RESTful==0.3.9
flatbuffers==23.1.4
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
invoke==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
paramiko==3.0.0
Pillow==9.4.0
prettytable==3.6.0
protobuf==3.19.6
py-edamam==0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
